refactor(content): align admin api with content lifecycle use cases

+++ src/hosts/CommercialNews.Api/Api/Admin/Contracts/Content/Categories/Responses/RestoreCategoryResponse.cs
@@ -4,7 +4,7 @@ namespace CommercialNews.Api.Api.Admin.Contracts.Content.Categories.Responses
     {
         public long CategoryId { get; init; }
         public bool IsDeleted { get; init; }
-        public int Version { get; init; }
+        public long Version { get; init; }
         public DateTime UpdatedAt { get; init; }
     }
 }
