+++ src/hosts/CommercialNews.Api/Api/Admin/Contracts/Content/Tags/Requests/UpdateTagRequest.cs
@@ -5,6 +5,6 @@ namespace CommercialNews.Api.Api.Admin.Contracts.Content.Tags.Requests
         public string Name { get; init; } = string.Empty;
         public string? Description { get; init; }
         public bool IsActive { get; init; }
-        public int ExpectedVersion { get; init; }
+        public long ExpectedVersion { get; init; }
     }
 }
